Move set_max_connections into util module
